updated lineitemprice
