+++ hamza-client/src/modules/common/components/line-item-unit-price/index.tsx
@@ -16,18 +16,18 @@ type LineItemUnitPriceProps = {
     item: Omit<ExtendedLineItem, 'beforeInsert'>;
     region: Region;
     style?: 'default' | 'tight';
+    currencyCode?: string;
 };
 
 const LineItemUnitPrice = ({
     item,
     region,
     style = 'default',
+    currencyCode,
 }: LineItemUnitPriceProps) => {
-    const { perferredCurrency } = useCustomerProfileStore();
-
     const unitPrice = item.variant.prices
         ? (item.variant as CalculatedVariant).prices.find(
-              (p) => p.currency_code == perferredCurrency
+              (p) => p.currency_code == currencyCode
           )?.amount ?? 0
         : (item.variant as CalculatedVariant)?.original_price ?? 0;
     const price = unitPrice * item.quantity;
